更新6.3.x

+++ selenium/che-selenium-core/src/main/java/org/eclipse/che/selenium/core/inject/SeleniumTestHandler.java
@@ -41,8 +41,10 @@ import java.util.concurrent.atomic.AtomicBoolean;
 import java.util.stream.Stream;
 import javax.annotation.PreDestroy;
 import javax.validation.constraints.NotNull;
+import org.eclipse.che.commons.json.JsonParseException;
 import org.eclipse.che.commons.lang.NameGenerator;
 import org.eclipse.che.selenium.core.SeleniumWebDriver;
+import org.eclipse.che.selenium.core.TestGroup;
 import org.eclipse.che.selenium.core.client.TestGitHubServiceClient;
 import org.eclipse.che.selenium.core.constant.TestBrowser;
 import org.eclipse.che.selenium.core.organization.InjectTestOrganization;
@@ -50,7 +52,10 @@ import org.eclipse.che.selenium.core.pageobject.InjectPageObject;
 import org.eclipse.che.selenium.core.pageobject.PageObjectsInjector;
 import org.eclipse.che.selenium.core.user.InjectTestUser;
 import org.eclipse.che.selenium.core.user.TestUser;
+import org.eclipse.che.selenium.core.webdriver.log.WebDriverLogsReaderFactory;
 import org.eclipse.che.selenium.core.workspace.InjectTestWorkspace;
+import org.eclipse.che.selenium.core.workspace.TestWorkspace;
+import org.eclipse.che.selenium.core.workspace.TestWorkspaceLogsReader;
 import org.eclipse.che.selenium.core.workspace.TestWorkspaceProvider;
 import org.openqa.selenium.OutputType;
 import org.openqa.selenium.WebDriver;
@@ -84,8 +89,20 @@ public abstract class SeleniumTestHandler
   private static final AtomicBoolean isCleanUpCompleted = new AtomicBoolean();
 
   @Inject
-  @Named("tests.screenshot_dir")
-  private String screenshotDir;
+  @Named("tests.screenshots_dir")
+  private String screenshotsDir;
+
+  @Inject
+  @Named("tests.htmldumps_dir")
+  private String htmldumpsDir;
+
+  @Inject
+  @Named("tests.webdriverlogs_dir")
+  private String webDriverLogsDir;
+
+  @Inject
+  @Named("tests.workspacelogs_dir")
+  private String workspaceLogsDir;
 
   @Inject private PageObjectsInjector pageObjectsInjector;
 
@@ -105,17 +122,24 @@ public abstract class SeleniumTestHandler
   @Named("sys.driver.version")
   private String webDriverVersion;
 
-  @Inject
+  @Inject(optional = true)
   @Named("github.username")
   private String gitHubUsername;
 
-  @Inject
+  @Inject(optional = true)
   @Named("github.password")
   private String gitHubPassword;
+
+  @Inject(optional = true)
+  @Named("sys.excludedGroups")
+  private String excludedGroups;
 
   @Inject private TestUser defaultTestUser;
   @Inject private TestWorkspaceProvider testWorkspaceProvider;
   @Inject private TestGitHubServiceClient gitHubClientService;
+  @Inject private TestWorkspaceLogsReader testWorkspaceLogsReader;
+  @Inject private SeleniumTestStatistics seleniumTestStatistics;
+  @Inject private WebDriverLogsReaderFactory webDriverLogsReaderFactory;
 
   private final Injector injector;
   private final Map<Long, Object> runningTests = new ConcurrentHashMap<>();
@@ -127,9 +151,15 @@ public abstract class SeleniumTestHandler
     getRuntime().addShutdownHook(new Thread(this::shutdown));
 
     revokeGithubOauthToken();
+    checkWebDriverSessionCreation();
   }
 
   private void revokeGithubOauthToken() {
+    // do not revoke if github tests are not being executed
+    if (excludedGroups == null || excludedGroups.contains(TestGroup.GITHUB)) {
+      return;
+    }
+
     try {
       gitHubClientService.deleteAllGrants(gitHubUsername, gitHubPassword);
     } catch (Exception e) {
@@ -138,40 +168,57 @@ public abstract class SeleniumTestHandler
   }
 
   @Override
-  public void onTestStart(ITestResult result) {}
+  public void onTestStart(ITestResult result) {
+    // count several invocations of method (e.g. in case of data provider is used)
+    String invocationLabel = "";
+    if (result.getMethod().getCurrentInvocationCount() > 0) {
+      invocationLabel = format(" (run %d)", result.getMethod().getCurrentInvocationCount() + 1);
+    }
+
+    LOG.info(
+        "Starting test #{} {}.{}{}. {}",
+        seleniumTestStatistics.hitStart(),
+        result.getTestClass().getRealClass().getSimpleName(),
+        result.getMethod().getMethodName(),
+        invocationLabel,
+        seleniumTestStatistics.toString());
+  }
 
   @Override
   public void onTestSuccess(ITestResult result) {
+    seleniumTestStatistics.hitPass();
     onTestFinish(result);
   }
 
   @Override
   public void onTestFailure(ITestResult result) {
+    seleniumTestStatistics.hitFail();
     onTestFinish(result);
   }
 
   @Override
   public void onTestSkipped(ITestResult result) {
+    seleniumTestStatistics.hitSkip();
     onTestFinish(result);
   }
 
   @Override
   public void onTestFailedButWithinSuccessPercentage(ITestResult result) {
+    seleniumTestStatistics.hitFail();
     onTestFinish(result);
   }
 
   @Override
-  public void onStart(ITestContext context) {
-    checkWebDriverSessionCreation();
-  }
+  public void onStart(ITestContext context) {}
 
   @Override
   public void onFinish(ITestContext context) {}
 
   @Override
   public void onStart(ISuite suite) {
-    runningTests.clear();
     suite.setParentInjector(injector);
+    LOG.info(
+        "Starting suite '{}' with {} test methods.", suite.getName(), suite.getAllMethods().size());
   }
 
   /** Check if webdriver session can be created without errors. */
@@ -248,22 +295,61 @@ public abstract class SeleniumTestHandler
   /** Is invoked when test or configuration is finished. */
   private void onTestFinish(ITestResult result) {
     if (result.getStatus() == ITestResult.FAILURE || result.getStatus() == ITestResult.SKIP) {
+      String invocationLabel = "";
+      if (result.getMethod().getCurrentInvocationCount() > 1) {
+        invocationLabel = format(" (run %d)", result.getMethod().getCurrentInvocationCount());
+      }
+
       if (result.getThrowable() != null) {
         LOG.error(
-            "Test {} method {} failed because {}",
-            result.getTestClass().getName(),
+            "Test {}.{}{} failed. Error: {}",
+            result.getTestClass().getRealClass().getSimpleName(),
             result.getMethod().getMethodName(),
+            invocationLabel,
             result.getThrowable().getLocalizedMessage());
         LOG.debug(result.getThrowable().getLocalizedMessage(), result.getThrowable());
       } else {
         LOG.error(
-            "Test {} method {} failed ",
-            result.getTestClass().getName(),
-            result.getMethod().getMethodName());
+            "Test {}.{}{} failed without exception.",
+            result.getTestClass().getRealClass().getSimpleName(),
+            result.getMethod().getMethodName(),
+            invocationLabel);
       }
+
       captureScreenshot(result);
       captureHtmlSource(result);
+      captureTestWorkspaceLogs(result);
+      storeWebDriverLogs(result);
     }
+  }
+
+  private void captureTestWorkspaceLogs(ITestResult result) {
+    Object testInstance = result.getInstance();
+    for (Field field : testInstance.getClass().getDeclaredFields()) {
+      field.setAccessible(true);
+
+      Object obj;
+      try {
+        obj = field.get(testInstance);
+      } catch (IllegalAccessException e) {
+        LOG.error(
+            "Field {} is inaccessible in {}.", field.getName(), testInstance.getClass().getName());
+        continue;
+      }
+
+      if (obj == null || !(obj instanceof TestWorkspace) || !isInjectedWorkspace(field)) {
+        continue;
+      }
+
+      Path pathToStoreWorkspaceLogs = Paths.get(workspaceLogsDir, getTestReference(result));
+      testWorkspaceLogsReader.read((TestWorkspace) obj, pathToStoreWorkspaceLogs);
+    }
+  }
+
+  private boolean isInjectedWorkspace(Field field) {
+    return field.isAnnotationPresent(com.google.inject.Inject.class)
+        || field.isAnnotationPresent(javax.inject.Inject.class)
+        || field.isAnnotationPresent(InjectTestWorkspace.class);
   }
 
   /** Releases resources by invoking methods annotated with {@link PreDestroy} */
@@ -278,7 +364,7 @@ public abstract class SeleniumTestHandler
         obj = field.get(testInstance);
       } catch (IllegalAccessException e) {
         LOG.error(
-            "Field {} is unaccessable in {}.", field.getName(), testInstance.getClass().getName());
+            "Field {} is inaccessible in {}.", field.getName(), testInstance.getClass().getName());
         continue;
       }
 
@@ -341,7 +427,7 @@ public abstract class SeleniumTestHandler
         obj = field.get(testInstance);
       } catch (IllegalAccessException e) {
         LOG.error(
-            "Field {} is unaccessable in {}.", field.getName(), testInstance.getClass().getName());
+            "Field {} is inaccessible in {}.", field.getName(), testInstance.getClass().getName());
         continue;
       }
 
@@ -369,16 +455,20 @@ public abstract class SeleniumTestHandler
   }
 
   private void captureScreenshotFromWindow(ITestResult result, SeleniumWebDriver webDriver) {
-    String testName = result.getTestClass().getName() + "." + result.getMethod().getMethodName();
-    String filename = NameGenerator.generate(testName + "_", 8) + ".png";
+    String testReference = getTestReference(result);
+    String filename = NameGenerator.generate(testReference + "_", 8) + ".png";
     try {
       byte[] data = webDriver.getScreenshotAs(OutputType.BYTES);
-      Path screenshot = Paths.get(screenshotDir, filename);
+      Path screenshot = Paths.get(screenshotsDir, filename);
       Files.createDirectories(screenshot.getParent());
       Files.copy(new ByteArrayInputStream(data), screenshot);
     } catch (WebDriverException | IOException e) {
-      LOG.error(format("Can't capture screenshot for test %s", testName), e);
+      LOG.error(format("Can't capture screenshot for test %s", testReference), e);
     }
+  }
+
+  private String getTestReference(ITestResult result) {
+    return result.getTestClass().getName() + "." + result.getMethod().getMethodName();
   }
 
   private void captureScreenshotsFromOpenedWindows(
@@ -392,17 +482,43 @@ public abstract class SeleniumTestHandler
             });
   }
 
+  private void storeWebDriverLogs(ITestResult result) {
+    Set<SeleniumWebDriver> webDrivers = new HashSet<>();
+    Object testInstance = result.getInstance();
+    collectInjectedWebDrivers(testInstance, webDrivers);
+    webDrivers.forEach(webDriver -> storeWebDriverLogs(result, webDriver));
+  }
+
+  private void storeWebDriverLogs(ITestResult result, SeleniumWebDriver webDriver) {
+    String testReference = getTestReference(result);
+
+    try {
+      String filename = NameGenerator.generate(testReference + "_", 4) + ".log";
+      Path webDriverLogsDirectory = Paths.get(webDriverLogsDir, filename);
+      Files.createDirectories(webDriverLogsDirectory.getParent());
+      Files.write(
+          webDriverLogsDirectory,
+          webDriverLogsReaderFactory
+              .create(webDriver)
+              .getAllLogs()
+              .getBytes(Charset.forName("UTF-8")),
+          StandardOpenOption.CREATE);
+    } catch (WebDriverException | IOException | JsonParseException e) {
+      LOG.error(format("Can't store web driver logs related to test %s.", testReference), e);
+    }
+  }
+
   private void dumpHtmlCodeFromTheCurrentPage(ITestResult result, SeleniumWebDriver webDriver) {
-    String testName = result.getTestClass().getName() + "." + result.getMethod().getMethodName();
-    String filename = NameGenerator.generate(testName + "_", 8) + ".html";
+    String testReference = getTestReference(result);
+    String filename = NameGenerator.generate(testReference + "_", 8) + ".html";
     try {
       String pageSource = webDriver.getPageSource();
-      Path dumpDirectory = Paths.get("target/htmldumps", filename);
+      Path dumpDirectory = Paths.get(htmldumpsDir, filename);
       Files.createDirectories(dumpDirectory.getParent());
       Files.write(
           dumpDirectory, pageSource.getBytes(Charset.forName("UTF-8")), StandardOpenOption.CREATE);
     } catch (WebDriverException | IOException e) {
-      LOG.error(format("Can't dump of html source for test %s", testName), e);
+      LOG.error(format("Can't dump of html source for test %s", testReference), e);
     }
   }
 
@@ -423,7 +539,7 @@ public abstract class SeleniumTestHandler
     }
 
     if (defaultTestUser != null) {
-      defaultTestUser.delete();
+      defaultTestUser.cleanUp();
     }
 
     isCleanUpCompleted.set(true);
